refactor: Phase1 弹幕合并器模块化（Wave1 A/B/C + Agent-D 集成）
