+++ registry/lib/components/video/danmaku/merger/options.ts
@@ -1,0 +1,11 @@
+import { defineOptionsMetadata, OptionsOfMetadata } from '@/components/define'
+
+/** 弹幕合并器组件设置（Phase 3 扩展维护操作区） */
+export const options = defineOptionsMetadata({
+  showMaintenanceActions: {
+    defaultValue: true,
+    displayName: '显示维护操作',
+  },
+})
+
+export type MergerOptions = OptionsOfMetadata<typeof options>
